feat: 게시글 삭제 시 댓글 작성자가 모두 같은 경우에만 삭제하는 기능 추가
삭제하기 전에 post의 writer와 post_id에 해당하는 comment 중에 writer가 다른 comment의 개수를 구하고
개수가 0일 때만 삭제하도록 구현했습니다.
게시글 삭제 후에 댓글도 모두 삭제하는 기능을 추가했습니다.

+++ src/main/java/kr/codesqaud/cafe/board/service/BoardService.java
@@ -3,6 +3,8 @@ package kr.codesqaud.cafe.board.service;
 import kr.codesqaud.cafe.board.dto.PostResponse;
 import kr.codesqaud.cafe.board.dto.PostWriteForm;
 import kr.codesqaud.cafe.board.repository.BoardJdbcRepository;
+import kr.codesqaud.cafe.board.repository.CommentJdbcRepository;
+import kr.codesqaud.cafe.exception.ForbiddenException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,9 +13,11 @@ import java.util.stream.Collectors;
 @Service
 public class BoardService {
     private final BoardJdbcRepository boardJdbcRepository;
+    private final CommentJdbcRepository commentJdbcRepository;
 
-    public BoardService(BoardJdbcRepository boardJdbcRepository) {
+    public BoardService(BoardJdbcRepository boardJdbcRepository, CommentJdbcRepository commentJdbcRepository) {
         this.boardJdbcRepository = boardJdbcRepository;
+        this.commentJdbcRepository = commentJdbcRepository;
     }
 
     public void write(PostWriteForm postWriteForm) {
@@ -25,7 +29,12 @@ public class BoardService {
     }
 
     public void delete(Long postId) {
-        boardJdbcRepository.delete(postId);
+        if (commentJdbcRepository.getCommentCountByOtherWriter(postId) == 0) {
+            boardJdbcRepository.delete(postId);
+            commentJdbcRepository.deleteAllByPostId(postId);
+        } else {
+            throw new ForbiddenException("작성자와 다른 댓글 작성자가 존재하여 삭제할 수 없습니다.");
+        }
     }
 
     public PostResponse getPost(Long postId) {
